backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
python-multipart==0.0.6
mysql-connector-python==8.2.0
requests==2.31.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0